Fix unit tests for Go 1.7
The existing tests are fragile and racy.

Changes made:
* Converted t.Fatal to t.Error anytime it was called from a goroutine.
The godoc for the testing package explicitly forbids t.Fatal from being used
outside of the primary Test function.
* Fixed all uses of sync.WaitGroup. It is racy to call WaitGroup.Add inside
of the goroutine. Instead Add should be called *before* spawning the goroutine.
The goroutine is only responsible for calling WaitGroup.Done. Also, use defer
with Done and Wait so that it is easier to reason about the correctness.
* Added another "check" to errCheck for "server closed connection". The server
now returns that when the transport is abruptly killed instead of returning
io.EOF. A future commit should seriously consider refactoring errCheck to be
less dependent on error values not guaranteed by the Go compatibility rules.