Interrupted locking mechanism
The following commit introduces a lock around the Interrupted state.
Introduction of the lock helps prevent race conditions where accessing
internal state at the same time as the Server.handleInterrupt.

Ideally I would prefer to handle the lock outside of this, but because
an internal state is made public via the struct and we have logic
that depends on that, we're a bit limited on how to fix this.